More dialogue additions
Added a button press to trigger the dialogue with NPC's when the player is within range. Once the initial converation is concluded, a second dialogue conversation can take place - meaning no repeated dialogue with the player. Added in a piece of UI to the NPC, doesn't do anything now but I will return to this in the future.