Multiple Major Changes
Firstly, the third person camera (TPCamera in script) is now complete and working as intended.
Secondly, I have removed the player controller script and instead made the player movement script a monobehaviour that attaches to the player object - whereas before the player controller was attached and all it was doing was calling a function on the player movement script which then moved the player. This just seemed weird and so I have changed the player movement script to to be the main component, as well as removed the need to use the rigidbody to move the player.
However, when the player collides withe NPC, the player start to spin on its own - even as the player is moved, and will continue to rotate. Looking into fixing this problem now.
Lastly, added the ability to lock the cursor and hide it when playing. The cursor returns when in dialogue, however the camera still rotates.