Major UI changes
Changed how the dialogue box looks and removed the animator so now the box just pops up. Fixed up some of the code to make it easier to read.
Added two images: a checkmark to indicate that the conversation is about the end, and a simple arrow to allow the player to move to the next line.