Stat system started
Added functionality for the player to get stat boosts when equiping armour and weapons using JSON. These json files contain the name of the item, its descripition, store worth and if it is a default item.

The player now takes damage (on demand for testing) and the modifiers of the armour is taken into consideration before reducing health.